Possibility to use install.php With this change it's possible to not remove install.php on container startup. This is usefull if there is no DB.

+++ 2.11/docker-entrypoint.d/20-web-console.sh
@@ -49,7 +49,7 @@ chown -R $APACHE_RUN_USER: $OCS_VARLIB_DIR
 chown -R $APACHE_RUN_USER: $OCS_WEBCONSOLE_DIR
 
 # rm install.php
-if [ -f $OCS_WEBCONSOLE_DIR/ocsreports/install.php ]; then
+if [ -f $OCS_WEBCONSOLE_DIR/ocsreports/install.php ] && [ $OCS_INSTALL_ENABLED -eq "0" ]; then
     rm $OCS_WEBCONSOLE_DIR/ocsreports/install.php
 fi
